refactor(web): extract query hooks to queries.ts in storage forms
Renames logical-volume-form/data.ts to queries.ts and creates a matching
queries.ts in partition-form, extracting the route param resolution hooks
from Form.tsx.

The "queries" name better describes what these modules do: they compose
route params with config model queries to derive the data forms need.
