+++ web/src/components/storage/logical-volume-form/Form.tsx
@@ -53,7 +53,7 @@ import {
   useUnusedLogicalVolumes,
   useInitialLogicalVolumeConfig,
   useUnusedMountPoints,
-} from "./data";
+} from "./queries";
 import {
   defaultOptions,
   validate,
